[SPOT-292][REFACTOR] 게시글 관련 DTO 요청/응답 분류
