+++ src/main/java/com/example/spot/post/application/PostCommandServiceImpl.java
@@ -20,13 +20,13 @@ import com.example.spot.post.domain.PostRepository;
 import com.example.spot.comment.presentation.dto.CommentCreateRequest;
 import com.example.spot.comment.presentation.dto.CommentCreateResponse;
 import com.example.spot.comment.presentation.dto.CommentLikeResponse;
-import com.example.spot.post.presentation.dto.PostCreateRequest;
-import com.example.spot.post.presentation.dto.PostCreateResponse;
-import com.example.spot.post.presentation.dto.PostLikeResponse;
-import com.example.spot.post.presentation.dto.PostUpdateRequest;
-import com.example.spot.post.presentation.dto.ScrapAllDeleteRequest;
-import com.example.spot.post.presentation.dto.ScrapPostResponse;
-import com.example.spot.post.presentation.dto.ScrapsPostDeleteResponse;
+import com.example.spot.post.presentation.dto.request.PostCreateRequest;
+import com.example.spot.post.presentation.dto.response.PostCreateResponse;
+import com.example.spot.post.presentation.dto.response.PostLikeResponse;
+import com.example.spot.post.presentation.dto.request.PostUpdateRequest;
+import com.example.spot.post.presentation.dto.request.ScrapAllDeleteRequest;
+import com.example.spot.post.presentation.dto.response.ScrapPostResponse;
+import com.example.spot.post.presentation.dto.response.ScrapsPostDeleteResponse;
 import com.example.spot.common.application.s3.S3ImageService;
 import com.example.spot.common.presentation.dto.util.response.ImageResponse.ImageUploadResponse;
 import com.example.spot.common.presentation.dto.util.response.ImageResponse.Images;
