+++ src/main/java/com/example/spot/post/application/PostQueryService.java
@@ -1,11 +1,11 @@
 package com.example.spot.post.application;
 
 import com.example.spot.comment.presentation.dto.CommentResponse;
-import com.example.spot.post.presentation.controller.PostAnnouncementResponse;
-import com.example.spot.post.presentation.dto.PostBest5Response;
-import com.example.spot.post.presentation.dto.PostPagingResponse;
-import com.example.spot.post.presentation.dto.PostRepresentativeResponse;
-import com.example.spot.post.presentation.dto.PostSingleResponse;
+import com.example.spot.post.presentation.dto.response.PostAnnouncementResponse;
+import com.example.spot.post.presentation.dto.response.PostBest5Response;
+import com.example.spot.post.presentation.dto.response.PostPagingResponse;
+import com.example.spot.post.presentation.dto.response.PostRepresentativeResponse;
+import com.example.spot.post.presentation.dto.response.PostSingleResponse;
 import org.springframework.data.domain.Pageable;
 
 public interface PostQueryService {
