+++ src/main/java/com/example/spot/post/application/PostQueryServiceImpl.java
@@ -13,14 +13,14 @@ import com.example.spot.report.domain.PostReportRepository;
 import com.example.spot.post.domain.PostRepository;
 import com.example.spot.comment.presentation.dto.CommentDetailResponse;
 import com.example.spot.comment.presentation.dto.CommentResponse;
-import com.example.spot.post.presentation.controller.PostAnnouncementResponse;
-import com.example.spot.post.presentation.dto.PostBest5DetailResponse;
-import com.example.spot.post.presentation.dto.PostBest5Response;
-import com.example.spot.post.presentation.dto.PostPagingDetailResponse;
-import com.example.spot.post.presentation.dto.PostPagingResponse;
-import com.example.spot.post.presentation.dto.PostRepresentativeDetailResponse;
-import com.example.spot.post.presentation.dto.PostRepresentativeResponse;
-import com.example.spot.post.presentation.dto.PostSingleResponse;
+import com.example.spot.post.presentation.dto.response.PostAnnouncementResponse;
+import com.example.spot.post.presentation.dto.response.PostBest5DetailResponse;
+import com.example.spot.post.presentation.dto.response.PostBest5Response;
+import com.example.spot.post.presentation.dto.response.PostPagingDetailResponse;
+import com.example.spot.post.presentation.dto.response.PostPagingResponse;
+import com.example.spot.post.presentation.dto.response.PostRepresentativeDetailResponse;
+import com.example.spot.post.presentation.dto.response.PostRepresentativeResponse;
+import com.example.spot.post.presentation.dto.response.PostSingleResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.data.domain.Page;
