+++ src/main/java/com/example/spot/post/presentation/controller/PostController.java
@@ -5,22 +5,23 @@ import com.example.spot.common.api.code.status.SuccessStatus;
 import com.example.spot.common.security.utils.SecurityUtils;
 import com.example.spot.post.application.PostCommandService;
 import com.example.spot.post.application.PostQueryService;
+import com.example.spot.post.presentation.dto.response.PostAnnouncementResponse;
 import com.example.spot.post.presentation.validator.ExistPost;
 import com.example.spot.comment.presentation.dto.CommentCreateRequest;
 import com.example.spot.comment.presentation.dto.CommentCreateResponse;
 import com.example.spot.comment.presentation.dto.CommentLikeResponse;
 import com.example.spot.comment.presentation.dto.CommentResponse;
-import com.example.spot.post.presentation.dto.PostBest5Response;
-import com.example.spot.post.presentation.dto.PostCreateRequest;
-import com.example.spot.post.presentation.dto.PostCreateResponse;
-import com.example.spot.post.presentation.dto.PostLikeResponse;
-import com.example.spot.post.presentation.dto.PostPagingResponse;
-import com.example.spot.post.presentation.dto.PostRepresentativeResponse;
-import com.example.spot.post.presentation.dto.PostSingleResponse;
-import com.example.spot.post.presentation.dto.PostUpdateRequest;
-import com.example.spot.post.presentation.dto.ScrapAllDeleteRequest;
-import com.example.spot.post.presentation.dto.ScrapPostResponse;
-import com.example.spot.post.presentation.dto.ScrapsPostDeleteResponse;
+import com.example.spot.post.presentation.dto.response.PostBest5Response;
+import com.example.spot.post.presentation.dto.request.PostCreateRequest;
+import com.example.spot.post.presentation.dto.response.PostCreateResponse;
+import com.example.spot.post.presentation.dto.response.PostLikeResponse;
+import com.example.spot.post.presentation.dto.response.PostPagingResponse;
+import com.example.spot.post.presentation.dto.response.PostRepresentativeResponse;
+import com.example.spot.post.presentation.dto.response.PostSingleResponse;
+import com.example.spot.post.presentation.dto.request.PostUpdateRequest;
+import com.example.spot.post.presentation.dto.request.ScrapAllDeleteRequest;
+import com.example.spot.post.presentation.dto.response.ScrapPostResponse;
+import com.example.spot.post.presentation.dto.response.ScrapsPostDeleteResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.media.Schema;
