+++ src/main/java/com/example/spot/post/presentation/dto/request/PostUpdateRequest.java
@@ -1,8 +1,7 @@
-package com.example.spot.post.presentation.dto;
+package com.example.spot.post.presentation.dto.request;
 
 import com.example.spot.post.domain.enums.Board;
 import io.swagger.v3.oas.annotations.media.Schema;
-import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -15,29 +14,30 @@ import org.springframework.web.multipart.MultipartFile;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class PostCreateRequest {
+public class PostUpdateRequest {
 
     @Schema(description = "게시글 제목입니다.",
             format = "string")
-    @NotNull
     private String title;
 
     @Schema(description = "게시글 내용입니다.",
             format = "string")
-    @NotNull
     private String content;
 
-    @Schema(
-            description = "게시글 타입입니다. " +
-                    "PASS_EXPERIENCE, INFORMATION_SHARING, COUNSELING, JOB_TALK, FREE_TALK, SPOT_ANNOUNCEMENT 중에 작성해주세요.",
-            example = "PASS_EXPERIENCE"
-    )
-    private Board type;
-
     @Schema(description = "익명 여부", example = "false")
-    private boolean anonymous;
+    private boolean isAnonymous;
 
-    @Schema(description = "이미지 파일", type = "string", format = "binary")
+    @Schema(
+            description = "게시글 타입입니다. 아래와 같이 작성해주세요.",
+            allowableValues = {"ALL", "PASS_EXPERIENCE", "INFORMATION_SHARING", "COUNSELING", "JOB_TALK", "FREE_TALK", "SPOT_ANNOUNCEMENT"}
+    )
+    private String type;
+
     private MultipartFile image;
 
+    private String existingImage;
+
+    public Board getType() {
+        return Board.findByValue(type);
+    }
 }
