+++ src/main/java/com/example/spot/post/presentation/dto/response/PostAnnouncementResponse.java
@@ -1,12 +1,10 @@
-package com.example.spot.post.presentation.controller;
+package com.example.spot.post.presentation.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 
 import java.util.List;
-
-import com.example.spot.post.presentation.dto.PostBest5DetailResponse;
 
 @Builder
 @Getter
