+++ src/main/java/com/example/spot/post/presentation/dto/response/PostHomeResponse.java
@@ -1,4 +1,4 @@
-package com.example.spot.post.presentation.dto;
+package com.example.spot.post.presentation.dto.response;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
